+++ src/ast/passes/semantic_analyser.h
@@ -99,9 +99,10 @@ private:
 
   SizedType *get_map_type(const Map &map);
   void assign_map_type(const Map &map, const SizedType &type);
-  void update_assign_map_type(const Map &map,
+  bool update_assign_map_type(const Map &map,
                               SizedType &type,
                               const SizedType &new_type);
+  bool update_expr_assigned_to_map(const SizedType &map_type, Expression *expr);
 
   void builtin_args_tracepoint(AttachPoint *attach_point, Builtin &builtin);
   ProbeType single_provider_type(void);